+++ Taylan/Singly-Linked-List/cpp/singly_linked_list.cpp
@@ -1,49 +1,6 @@
-#include <iostream>
+#include "singly_linked_list.h"
+
 namespace std{
-
-	class Node{
-	private:
-		int _data;
-		Node *_next;
-
-	public:
-		Node(){}
-		// setters and getters
-		void setData(int Data){ _data = Data;}
-		void setNext(Node *Next){ 
-			if(Next == NULL){
-				_next = NULL;
-			}else{
-			_next = Next;
-			}
-		}
-		int Data(){return _data;}
-		Node *Next(){return _next;}
-	};
-
-	class LinkedList{
-	private:
-		Node *head;
-	public:
-		LinkedList(){ head = NULL;}//init
-		/*TODO: Tail will be added. */
-		void insert_back(int data);//it will append the next value to the
-									//front cell [x] -> [data]
-		void insert_front(int data);
-		bool isEmpty();
-		void init_list(int data);
-		void print_List();
-		int size();
-		void insert_into(int location, int data);
-		int value_at(int index);
-		int from_tail_value_at(int index);
-		int get_value_from_front();
-		int get_value_from_back();
-		int pop_front();
-		int pop_back();
-		void remove_at(int index);
-	};
-
 	void LinkedList::print_List(){
 		//init. the temporary pointer, so that we cannot lose
 		//original head pointer of the list.
@@ -228,34 +185,25 @@ namespace std{
 			Node *p = tmp->Next()->Next();
 			delete tmp->Next();
 			tmp->setNext(p);
-			
-
 		}
-
 	}
+	/*Reversing the linked list and printing the reversed order.*/
+	Node * LinkedList::reverse(){
+		Node *tmp = head;
+		Node *sec = new Node();
+		Node *reverselist = NULL;
 
-}
-
-int main(){
-	//Creating a list
-	std::LinkedList list;
-	/*Since this part for the testing changes do not matter.*/
-	//Initilizing it with 5
-	list.init_list(5);
-	list.insert_back(6);
-	list.insert_front(4);
-	list.insert_into(1, 7);
-	list.print_List();
-	// std::cout<<'\n'<<list.from_tail_value_at(4);
-	// std::cout<<'\n'<<list.value_at(1);
-	// std::cout<<'\n'<<list.size();//output2.
-	// std::cout<<'\n'<<list.isEmpty();
-	// std::cout<<'\n'<<list.get_value_from_front();
-	// std::cout<<'\n'<<list.get_value_from_back();
-	// std::cout<<'\n'<<list.pop_front()<<'\n';
-	// list.print_List();
-	// std::cout<<'\n'<<list.pop_back()<<'\n';
-	std::cout<<'\n';
-	list.remove_at(7);
-	list.print_List();
+		/*Reversing the list*/
+		while(tmp != NULL){
+			sec = tmp->Next();
+			tmp->setNext(reverselist);
+			reverselist = tmp;
+			tmp = sec;
+		}
+		while(reverselist){
+			cout<<reverselist->Data()<<" > ";
+			reverselist = reverselist->Next();
+		}
+		return reverselist;
+	}
 }
